Data.Bytes.Put

Description

This module generalizes the binary PutM and cereal PutM monads in an ad hoc fashion to permit code to be written that is compatible across them.

Moreover, this class permits code to be written to be portable over various monad transformers applied to these as base monads.

Documentation

class (Applicative m, Monad m) => MonadPut m where

Minimal complete definition

Nothing

Methods

putWord8 :: Word8 -> m ()

Efficiently write a byte into the output buffer

default putWord8 :: (m ~ t n, MonadTrans t, MonadPut n) => Word8 -> m ()

putByteString :: ByteString -> m ()

An efficient primitive to write a strict ByteString into the output buffer.

In binary this flushes the current buffer, and writes the argument into a new chunk.

default putByteString :: (m ~ t n, MonadTrans t, MonadPut n) => ByteString -> m ()

putLazyByteString :: ByteString -> m ()

Write a lazy ByteString efficiently.

With binary, this simply appends the chunks to the output buffer

default putLazyByteString :: (m ~ t n, MonadTrans t, MonadPut n) => ByteString -> m ()

flush :: m ()

Pop the ByteString we have constructed so far, if any, yielding a new chunk in the result ByteString.

If we're building a strict ByteString with cereal then this does nothing.

default flush :: (m ~ t n, MonadTrans t, MonadPut n) => m ()

putWord16le :: Word16 -> m ()

Write a Word16 in little endian format

default putWord16le :: (m ~ t n, MonadTrans t, MonadPut n) => Word16 -> m ()

putWord16be :: Word16 -> m ()

Write a Word16 in big endian format

default putWord16be :: (m ~ t n, MonadTrans t, MonadPut n) => Word16 -> m ()

putWord16host :: Word16 -> m ()

O(1). Write a Word16 in native host order and host endianness. For portability issues see putWordhost.

default putWord16host :: (m ~ t n, MonadTrans t, MonadPut n) => Word16 -> m ()

putWord32le :: Word32 -> m ()

Write a Word32 in little endian format

default putWord32le :: (m ~ t n, MonadTrans t, MonadPut n) => Word32 -> m ()

putWord32be :: Word32 -> m ()

Write a Word32 in big endian format

default putWord32be :: (m ~ t n, MonadTrans t, MonadPut n) => Word32 -> m ()

putWord32host :: Word32 -> m ()

O(1). Write a Word32 in native host order and host endianness. For portability issues see putWordhost.

default putWord32host :: (m ~ t n, MonadTrans t, MonadPut n) => Word32 -> m ()

putWord64le :: Word64 -> m ()

Write a Word64 in little endian format

default putWord64le :: (m ~ t n, MonadTrans t, MonadPut n) => Word64 -> m ()

putWord64be :: Word64 -> m ()

Write a Word64 in big endian format

default putWord64be :: (m ~ t n, MonadTrans t, MonadPut n) => Word64 -> m ()

putWord64host :: Word64 -> m ()

O(1). Write a Word64 in native host order and host endianness. For portability issues see putWordhost.

default putWord64host :: (m ~ t n, MonadTrans t, MonadPut n) => Word64 -> m ()

putWordhost :: Word -> m ()

O(1). Write a single native machine word. The word is written in host order, host endian form, for the machine you're on. On a 64 bit machine the Word is an 8 byte value, on a 32 bit machine, 4 bytes. Values written this way are not portable to different endian or word sized machines, without conversion.

default putWordhost :: (m ~ t n, MonadTrans t, MonadPut n) => Word -> m ()

runPutL :: Put -> ByteString

Put a value into a lazy ByteString using runPut.

runPutS :: Put -> ByteString

Put a value into a strict ByteString using runPut.