Data.Bytes.Serial

Description

This module generalizes the binary PutM and cereal PutM monads in an ad hoc fashion to permit code to be written that is compatible across them.

Moreover, this class permits code to be written to be portable over various monad transformers applied to these as base monads.

Documentation

class Serial a where

Methods

serialize :: MonadPut m => a -> m ()

deserialize :: MonadGet m => m a

class GSerial f where

Used internally to provide generic serialization

Methods

gserialize :: MonadPut m => f a -> m ()

gdeserialize :: MonadGet m => m (f a)

class Serial1 f where

Methods

serializeWith :: MonadPut m => (a -> m ()) -> f a -> m ()

deserializeWith :: MonadGet m => m a -> m (f a)

serialize1 :: (MonadPut m, Serial1 f, Serial a) => f a -> m ()

deserialize1 :: (MonadGet m, Serial1 f, Serial a) => m (f a)

class GSerial1 f where

Used internally to provide generic serialization

Methods

gserializeWith :: MonadPut m => (a -> m ()) -> f a -> m ()

gdeserializeWith :: MonadGet m => m a -> m (f a)