Data.Bytes.Get

Description

This module generalizes the binary Get and cereal Get monads in an ad hoc fashion to permit code to be written that is compatible across them.

Moreover, this class permits code to be written to be portable over various monad transformers applied to these as base monads.

Documentation

class (forall a b. Coercible a b => Coercible (m a) (m b), Integral (Remaining m), MonadFail m, Applicative m) => MonadGet m where

Minimal complete definition

lookAhead, lookAheadM, lookAheadE

Associated Types

type Remaining m :: *

An Integral number type used for unchecked skips and counting.

type Bytes m :: *

The underlying ByteString type used by this instance

Methods

skip :: Int -> m ()

Skip ahead n bytes. Fails if fewer than n bytes are available.

default skip :: (MonadTrans t, MonadGet n, m ~ t n) => Int -> m ()

ensure :: Int -> m ByteString

If at least n bytes are available return at least that much of the current input. Otherwise fail.

default ensure :: (MonadTrans t, MonadGet n, m ~ t n) => Int -> m ByteString

lookAhead :: m a -> m a

Run ga, but return without consuming its input. Fails if ga fails.

lookAheadM :: m (Maybe a) -> m (Maybe a)

Like lookAhead, but consume the input if gma returns 'Just _'. Fails if gma fails.

lookAheadE :: m (Either a b) -> m (Either a b)

Like lookAhead, but consume the input if gea returns 'Right _'. Fails if gea fails.

getBytes :: Int -> m ByteString

Pull n bytes from the input, as a strict ByteString.

default getBytes :: (MonadTrans t, MonadGet n, m ~ t n) => Int -> m ByteString

remaining :: m (Remaining m)

Get the number of remaining unparsed bytes. Useful for checking whether all input has been consumed. Note that this forces the rest of the input.

default remaining :: (MonadTrans t, MonadGet n, m ~ t n, Remaining m ~ Remaining n) => m (Remaining m)

isEmpty :: m Bool

Test whether all input has been consumed, i.e. there are no remaining unparsed bytes.

default isEmpty :: (MonadTrans t, MonadGet n, m ~ t n) => m Bool

getWord8 :: m Word8

Read a Word8 from the monad state

default getWord8 :: (MonadTrans t, MonadGet n, m ~ t n) => m Word8

getByteString :: Int -> m ByteString

An efficient get method for strict ByteStrings. Fails if fewer than n bytes are left in the input.

default getByteString :: (MonadTrans t, MonadGet n, m ~ t n) => Int -> m ByteString

getLazyByteString :: Int64 -> m ByteString

An efficient get method for lazy ByteStrings. Does not fail if fewer than n bytes are left in the input.

default getLazyByteString :: (MonadTrans t, MonadGet n, m ~ t n) => Int64 -> m ByteString

getWord16be :: m Word16

Read a Word16 in big endian format

default getWord16be :: (MonadTrans t, MonadGet n, m ~ t n) => m Word16

getWord16le :: m Word16

Read a Word16 in little endian format

default getWord16le :: (MonadTrans t, MonadGet n, m ~ t n) => m Word16

getWord16host :: m Word16

O(1). Read a 2 byte Word16 in native host order and host endianness.

default getWord16host :: (MonadTrans t, MonadGet n, m ~ t n) => m Word16

getWord32be :: m Word32

Read a Word32 in big endian format

default getWord32be :: (MonadTrans t, MonadGet n, m ~ t n) => m Word32

getWord32le :: m Word32

Read a Word32 in little endian format

default getWord32le :: (MonadTrans t, MonadGet n, m ~ t n) => m Word32

getWord32host :: m Word32

O(1). Read a Word32 in native host order and host endianness.

default getWord32host :: (MonadTrans t, MonadGet n, m ~ t n) => m Word32

getWord64be :: m Word64

Read a Word64 in big endian format

default getWord64be :: (MonadTrans t, MonadGet n, m ~ t n) => m Word64

getWord64le :: m Word64

Read a Word64 in little endian format

default getWord64le :: (MonadTrans t, MonadGet n, m ~ t n) => m Word64

getWord64host :: m Word64

O(1). Read a Word64 in native host order and host endianness.

default getWord64host :: (MonadTrans t, MonadGet n, m ~ t n) => m Word64

getWordhost :: m Word

O(1). Read a single native machine word. The word is read in host order, host endian form, for the machine you're on. On a 64 bit machine the Word is an 8 byte value, on a 32 bit machine, 4 bytes.

default getWordhost :: (MonadTrans t, MonadGet n, m ~ t n) => m Word

runGetL :: Get a -> ByteString -> a

Get something from a lazy ByteString using runGet.

runGetS :: Get a -> ByteString -> Either String a

Get something from a strict ByteString using runGet.