Data.Bytes.Serial

Description

This module contains two main classes, each providing methods to serialize and deserialize types. Serial is the primary class, to be used for the canonical way to serialize a specific type. SerialEndian is used to provide endian-specific methods for serializing a type.

Serialization

class Serial a where

Methods to serialize and deserialize type a to a binary representation

Instances provided here for fixed-with Integers and Words are big endian. Instances for strict and lazy bytestrings store also the length of bytestring big endian. Instances for Word and Int are host endian as they are machine-specific types.

Methods

serialize :: MonadPut m => a -> m ()

serialize :: (MonadPut m, GSerial (Rep a), Generic a) => a -> m ()

deserialize :: MonadGet m => m a

deserialize :: (MonadGet m, GSerial (Rep a), Generic a) => m a

Specifying endianness

class SerialEndian a where

Methods to serialize and deserialize type a to a big and little endian binary representations. Methods suffixed with "host" are automatically defined to use equal the methods corresponding to the current machine's native endianness, but they can be overridden.

Methods

serializeBE :: MonadPut m => a -> m ()

serializeBE :: (MonadPut m, GSerialEndian (Rep a), Generic a) => a -> m ()

deserializeBE :: MonadGet m => m a

deserializeBE :: (MonadGet m, GSerialEndian (Rep a), Generic a) => m a

serializeLE :: MonadPut m => a -> m ()

serializeLE :: (MonadPut m, GSerialEndian (Rep a), Generic a) => a -> m ()

deserializeLE :: MonadGet m => m a

deserializeLE :: (MonadGet m, GSerialEndian (Rep a), Generic a) => m a

serializeHost :: MonadPut m => a -> m ()

deserializeHost :: MonadGet m => m a

Higher-order

These classes provide us with the ability to serialize containers that need polymorphic recursion.

class Serial1 f where

Methods

serializeWith :: MonadPut m => (a -> m ()) -> f a -> m ()

serializeWith :: (MonadPut m, GSerial1 (Rep1 f), Generic1 f) => (a -> m ()) -> f a -> m ()

deserializeWith :: MonadGet m => m a -> m (f a)

deserializeWith :: (MonadGet m, GSerial1 (Rep1 f), Generic1 f) => m a -> m (f a)

serialize1 :: (MonadPut m, Serial1 f, Serial a) => f a -> m ()

deserialize1 :: (MonadGet m, Serial1 f, Serial a) => m (f a)

class Serial2 f where

Minimal complete definition

serializeWith2, deserializeWith2

Methods

serializeWith2 :: MonadPut m => (a -> m ()) -> (b -> m ()) -> f a b -> m ()

deserializeWith2 :: MonadGet m => m a -> m b -> m (f a b)

serialize2 :: (MonadPut m, Serial2 f, Serial a, Serial b) => f a b -> m ()

deserialize2 :: (MonadGet m, Serial2 f, Serial a, Serial b) => m (f a b)

Storable

store :: (MonadPut m, Storable a) => a -> m ()

serialize any Storable in a host-specific format.

restore :: forall m a. (MonadGet m, Storable a) => m a

deserialize any Storable in a host-specific format.

Generics

You probably will never need to care that these exist except they provide us with default definitions for Serial and SerialEndian

class GSerial f where

Used internally to provide generic serialization

Minimal complete definition

gserialize, gdeserialize

Methods

gserialize :: MonadPut m => f a -> m ()

gdeserialize :: MonadGet m => m (f a)

class GSerialEndian f where

Used internally to provide generic big-endian serialization

Methods

gserializeBE :: MonadPut m => f a -> m ()

gserializeBE :: (MonadPut m, GSerial f) => f a -> m ()

gdeserializeBE :: MonadGet m => m (f a)

gdeserializeBE :: (MonadGet m, GSerial f) => m (f a)

gserializeLE :: MonadPut m => f a -> m ()

gserializeLE :: (MonadPut m, GSerial f) => f a -> m ()

gdeserializeLE :: MonadGet m => m (f a)

gdeserializeLE :: (MonadGet m, GSerial f) => m (f a)

class GSerial1 f where

Used internally to provide generic serialization

Minimal complete definition

gserializeWith, gdeserializeWith

Methods

gserializeWith :: MonadPut m => (a -> m ()) -> f a -> m ()

gdeserializeWith :: MonadGet m => m a -> m (f a)