Data.Bytes.Signed

Description

When one wants to think of an Int as a dumb bitstring, converting it to a Word avoids pesky complications with respect to sign extension.

Documentation

type family Unsigned i :: *

unsigned :: (Integral i, Num (Unsigned i)) => i -> Unsigned i

type family Signed i :: *

signed :: (Integral i, Num (Signed i)) => i -> Signed i